4. split and compation files.
